Move startImport action to redux
